backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==23.2
djoser==2.1.0
Pillow==9.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
psycopg2==2.9.7